Finished getting the particle system working again. Fixed several bugs when handling errors from Lua scripts. Attempt at better time slicing with the camera (doesn't seem to help stutters).

+++ TiVEPluginFramework/IBlockList.cs
@@ -9,6 +9,6 @@ namespace ProdigalSoftware.TiVEPluginFramework
         ushort this[string blockName] { get; }
 
         [UsedImplicitly]
-        Block this[int blockIndex] { get; }
+        Block this[ushort blockIndex] { get; }
     }
 }
